Funcs relocation
